requirements.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
psycopg2-binary>=2.8.5,<2.9.0
Pillow>=5.3.0,<5.4.0

flake8>=3.6.0<3.7.0
